refactor(desktop): split provider settings flows
