+++ apps/desktop/src/main/__tests__/provider-contract-source-helpers.ts
@@ -1,0 +1,66 @@
+import { readFile } from 'node:fs/promises';
+import { resolve } from 'node:path';
+
+const REPO_ROOT = resolve(import.meta.dirname, '../../../../..');
+const SETTINGS_ROOT = resolve(REPO_ROOT, 'apps', 'desktop', 'src', 'renderer', 'settings');
+
+export interface ProviderSettingsSources {
+  panel: string;
+  sheet: string;
+  catalog: string;
+  oauth: string;
+  display: string;
+  addForm: string;
+  detail: string;
+  shared: string;
+  combined: string;
+}
+
+const sourcePaths = {
+  panel: resolve(SETTINGS_ROOT, 'ProvidersPanel.tsx'),
+  sheet: resolve(SETTINGS_ROOT, 'provider-config-sheet.tsx'),
+  catalog: resolve(SETTINGS_ROOT, 'provider-catalog.tsx'),
+  oauth: resolve(SETTINGS_ROOT, 'provider-oauth-section.tsx'),
+  display: resolve(SETTINGS_ROOT, 'provider-display.tsx'),
+  addForm: resolve(SETTINGS_ROOT, 'provider-add-form.tsx'),
+  detail: resolve(SETTINGS_ROOT, 'provider-connection-detail.tsx'),
+  shared: resolve(SETTINGS_ROOT, 'provider-panel-shared.ts'),
+} as const;
+
+export async function readProviderSettingsSources(): Promise<ProviderSettingsSources> {
+  const [panel, sheet, catalog, oauth, display, addForm, detail, shared] = await Promise.all([
+    readFile(sourcePaths.panel, 'utf8'),
+    readFile(sourcePaths.sheet, 'utf8'),
+    readFile(sourcePaths.catalog, 'utf8'),
+    readFile(sourcePaths.oauth, 'utf8'),
+    readFile(sourcePaths.display, 'utf8'),
+    readFile(sourcePaths.addForm, 'utf8'),
+    readFile(sourcePaths.detail, 'utf8'),
+    readFile(sourcePaths.shared, 'utf8'),
+  ]);
+
+  return {
+    panel,
+    sheet,
+    catalog,
+    oauth,
+    display,
+    addForm,
+    detail,
+    shared,
+    combined: [
+      panel,
+      sheet,
+      catalog,
+      oauth,
+      display,
+      addForm,
+      detail,
+      shared,
+    ].join('\n'),
+  };
+}
+
+export async function readProviderSettingsCombinedSource(): Promise<string> {
+  return (await readProviderSettingsSources()).combined;
+}
