+++ apps/desktop/src/renderer/settings/ProvidersPanel.tsx
@@ -1,58 +1,32 @@
-import { useEffect, useMemo, useRef, useState, type KeyboardEvent as ReactKeyboardEvent, type ReactNode, type RefObject } from 'react';
-import { ChevronRight, X } from '@maka/ui/icons';
-import {
-  canPickDefaultModel,
-  canSaveDefaultModelChange,
-  nextRadioId,
-  selectableDefaultModelIds,
-  tabbableRadioId,
-} from './model-table-keyboard';
+import { useEffect, useMemo, useRef, useState } from 'react';
+import { ChevronRight } from '@maka/ui/icons';
 import {
   CATALOG_PROVIDER_TYPES,
   PROVIDER_DEFAULTS,
-  generalizedErrorMessageChinese,
-  redactSecrets,
-  validateSlug,
-  type ConnectionTestResult,
-  type CreateConnectionInput,
   type LlmConnection,
-  type ModelCatalogEntry,
-  type ModelDiscoveryResult,
-  type ModelInfo,
   type ProviderCategory,
   type ProviderType,
-  type SubscriptionAccountState,
-  type UpdateConnectionInput,
 } from '@maka/core';
 import {
   Button,
   PrimitiveTabs, PrimitiveTabsList, PrimitiveTabsTrigger,
   PrimitiveAccordion, PrimitiveAccordionItem, PrimitiveAccordionHeader, PrimitiveAccordionTrigger, PrimitiveAccordionPanel,
-  Item, ItemMedia, ItemContent, ItemTitle, ItemDescription, ItemActions,
-  FieldRoot, Label, FieldDescription,
-  Input, RelativeTime, Textarea, useToast, useModalA11y,
+  Item, ItemContent, ItemTitle, ItemActions,
+  useToast,
 } from '@maka/ui';
-import { formatRelativeTimestamp } from '@maka/core';
-import { PasswordInput } from './password-input';
-import { ProviderBrandMark } from './provider-brand-marks';
 import { chipStatusText, rollupForGroup } from './provider-connection-status';
-import { buildCatalogModelChoices, buildCatalogRecommendedDefaultModel } from '../model-catalog-choices';
+import { AddProviderForm } from './provider-add-form';
+import { ProviderCatalogCard } from './provider-catalog';
+import { ConnectionDetail } from './provider-connection-detail';
+import { ProviderConfigSheetOverlay } from './provider-config-sheet';
+import { ProviderLogo, providerDisplay } from './provider-display';
+import { ModelOAuthSection } from './provider-oauth-section';
+import { providerPanelActionErrorMessage, type ConnectionsBridge } from './provider-panel-shared';
 
-export interface ConnectionsBridge {
-  list(): Promise<LlmConnection[]>;
-  getDefault(): Promise<string | null>;
-  setDefault(slug: string | null): Promise<void>;
-  create(input: CreateConnectionInput): Promise<LlmConnection>;
-  update(slug: string, patch: UpdateConnectionInput): Promise<LlmConnection>;
-  delete(slug: string): Promise<void>;
-  test(slug: string, opts?: { model?: string }): Promise<ConnectionTestResult>;
-  fetchModels(slug: string): Promise<ModelDiscoveryResult>;
-  hasSecret(slug: string): Promise<boolean>;
-  subscribeEvents?(handler: () => void): () => void;
-}
+export type { ConnectionsBridge } from './provider-panel-shared';
+export { ProviderLogo, providerDisplay } from './provider-display';
 
 type CatalogTab = Extract<ProviderCategory, 'domestic' | 'overseas' | 'local' | 'oauth'>;
-type CredentialPresenceStatus = boolean | 'loading' | 'error';
 
 const CATALOG_TABS: Array<{ id: CatalogTab; label: string }> = [
   { id: 'domestic', label: '国内' },
@@ -60,42 +34,6 @@ const CATALOG_TABS: Array<{ id: CatalogTab; label: string }> = [
   { id: 'local', label: '本地' },
   { id: 'oauth', label: 'OAuth' },
 ];
-
-/**
- * "（5 分钟前拉取）" style suffix for the model-source label.
- * Delegates to the shared `@maka/core/relative-time` helper so the
- * format matches the rest of the app's relative-time surfaces.
- * Returns an empty string when no timestamp is available
- * (e.g. legacy connections from before `modelsFetchedAt` was
- * persisted by backend `94b482b`).
- */
-function formatFetchedAtSuffix(modelsFetchedAt: number | undefined): string {
-  if (modelsFetchedAt === undefined) return '';
-  return `（${formatRelativeTimestamp(modelsFetchedAt)}拉取）`;
-}
-
-function providerPanelActionErrorMessage(error: unknown): string {
-  return generalizedErrorMessageChinese(error, '模型连接服务暂时不可用，请稍后重试。');
-}
-
-function connectionTestFailureMessage(result: ConnectionTestResult, troubleshootingCopy: string): string {
-  const fallback = connectionTestFailureFallback(result, troubleshootingCopy);
-  if (!result.errorMessage) return fallback;
-  return generalizedErrorMessageChinese(new Error(result.errorMessage), fallback);
-}
-
-function connectionTestFailureFallback(result: ConnectionTestResult, troubleshootingCopy: string): string {
-  if (result.statusCode === 429) return '当前账号或模型服务触发速率限制，请稍后重试。';
-  if (result.errorClass === 'timeout') return '请求超时，请检查网络或代理后重试。';
-  if (result.errorClass === 'auth' || result.statusCode === 401 || result.statusCode === 403) {
-    return `鉴权失败，请确认 ${troubleshootingCopy} 后重试。`;
-  }
-  if (result.errorClass === 'provider_unavailable' || (result.statusCode !== undefined && result.statusCode >= 500)) {
-    return '模型服务暂时不可用，请稍后重试。';
-  }
-  if (result.errorClass === 'network') return '网络错误，请检查服务地址或代理设置后重试。';
-  return `检查 ${troubleshootingCopy} 后重试。`;
-}
 
 export function ProvidersPanel({ bridge }: { bridge: ConnectionsBridge }) {
   const [connections, setConnections] = useState<LlmConnection[]>([]);
@@ -440,2130 +378,4 @@ export function ProvidersPanel({ bridge }: { bridge: ConnectionsBridge }) {
       )}
     </div>
   );
-}
-
-/**
- * Modal overlay + sheet for the provider config sub-flow. Wraps
- * `useModalA11y` so:
- *  - Tab/Shift+Tab cycles focus inside the sheet (no leak to sidebar)
- *  - Initial focus lands on the first interactive element
- *  - Esc closes the sheet (matches the overlay click-to-close)
- *  - Focus restoration to the previously-focused element on close
- *
- * Without this hook the sheet had `role="dialog"` + `aria-modal="true"`
- * but no actual focus trap or keyboard-dismiss path — a screen reader
- * user couldn't navigate the sheet predictably.
- */
-function ProviderConfigSheetOverlay(props: { onClose(): void; children: ReactNode }) {
-  const dialogRef = useRef<HTMLElement>(null);
-  useModalA11y(dialogRef, props.onClose);
-  useProviderSheetBackgroundInert(dialogRef);
-  return (
-    <div className="providerConfigOverlay" role="presentation" onMouseDown={props.onClose}>
-      <section
-        ref={dialogRef as RefObject<HTMLDivElement>}
-        className="providerConfigSheet"
-        role="dialog"
-        aria-modal="true"
-        aria-label="模型供应商配置"
-        onMouseDown={(event) => event.stopPropagation()}
-      >
-        <Button
-          type="button"
-          variant="quiet"
-          size="icon-sm"
-          className="providerConfigSheetClose"
-          aria-label="关闭模型配置"
-          onClick={props.onClose}
-        >
-          <X strokeWidth={1.75} aria-hidden="true" />
-        </Button>
-        {props.children}
-      </section>
-    </div>
-  );
-}
-
-function useProviderSheetBackgroundInert(dialogRef: RefObject<HTMLElement | null>) {
-  useEffect(() => {
-    const dialog = dialogRef.current;
-    if (!dialog) return;
-    const surface = dialog.closest('.settingsSurface');
-    if (!(surface instanceof HTMLElement)) return;
-
-    const changed: Array<{
-      element: HTMLElement;
-      ariaHidden: string | null;
-      inert: boolean;
-      marker: string | null;
-    }> = [];
-    let current: HTMLElement | null = dialog;
-    while (current && current !== surface) {
-      const parent: HTMLElement | null = current.parentElement;
-      if (!parent) break;
-      for (const sibling of Array.from(parent.children)) {
-        if (!(sibling instanceof HTMLElement) || sibling === current || sibling.contains(dialog)) continue;
-        changed.push({
-          element: sibling,
-          ariaHidden: sibling.getAttribute('aria-hidden'),
-          inert: sibling.inert,
-          marker: sibling.getAttribute('data-provider-sheet-background-hidden'),
-        });
-        sibling.setAttribute('aria-hidden', 'true');
-        sibling.inert = true;
-        sibling.setAttribute('data-provider-sheet-background-hidden', 'true');
-      }
-      current = parent;
-    }
-
-    return () => {
-      for (const item of changed.reverse()) {
-        if (item.ariaHidden === null) item.element.removeAttribute('aria-hidden');
-        else item.element.setAttribute('aria-hidden', item.ariaHidden);
-        item.element.inert = item.inert;
-        if (item.marker === null) item.element.removeAttribute('data-provider-sheet-background-hidden');
-        else item.element.setAttribute('data-provider-sheet-background-hidden', item.marker);
-      }
-    };
-  }, [dialogRef]);
-}
-
-function ProviderCatalogCard(props: { type: ProviderType; count: number; onSelect(): void }) {
-  const defaults = PROVIDER_DEFAULTS[props.type];
-  const display = providerDisplay(props.type);
-  const disabled = defaults.status !== 'ready';
-  const disabledStatus = providerDisabledStatus(props.type);
-  const title = disabled ? providerDisabledTitle(props.type) : `添加 ${display.name}`;
-
-  if (disabled) {
-    return (
-      <Item
-        className="providerCatalogRow rounded-none"
-        data-provider={props.type}
-        data-status={disabledStatus}
-        data-disabled="true"
-        aria-label={providerDisabledAriaLabel(props.type, display.name)}
-        title={title}
-      >
-        <ItemMedia>
-          <ProviderLogo type={props.type} />
-        </ItemMedia>
-        <ItemContent>
-          <ItemTitle className="providerCatalogTitle">{display.name}</ItemTitle>
-          <ItemDescription className="providerCatalogDesc">{display.description}</ItemDescription>
-        </ItemContent>
-        <ItemActions>
-          <span className="providerCatalogBadge is-state" aria-hidden="true">
-            {disabledStatus === 'experimental' ? '实验' : '未开放'}
-          </span>
-        </ItemActions>
-      </Item>
-    );
-  }
-
-  return (
-    <Item
-      className="providerCatalogRow rounded-none"
-      data-provider={props.type}
-      data-status="ready"
-      aria-label={providerCatalogAriaLabel(display, props.count)}
-      title={title}
-      render={<button type="button" onClick={props.onSelect} />}
-    >
-      <ItemMedia>
-        <ProviderLogo type={props.type} />
-      </ItemMedia>
-      <ItemContent>
-        <ItemTitle className="providerCatalogTitle">{display.name}</ItemTitle>
-        <ItemDescription className="providerCatalogDesc">
-          {display.description}
-          {props.count > 0 && <span className="providerCatalogCount">已配置 {props.count} 个</span>}
-        </ItemDescription>
-      </ItemContent>
-      <ItemActions className="providerCatalogActions">
-        {display.badge && <span className="providerCatalogBadge">{display.badge}</span>}
-        <ChevronRight className="providerCatalogChevron" size={15} strokeWidth={2} aria-hidden="true" />
-      </ItemActions>
-    </Item>
-  );
-}
-
-function providerDisabledStatus(type: ProviderType): 'unavailable' | 'experimental' {
-  return isWiredOAuthProvider(type) ? 'experimental' : 'unavailable';
-}
-
-function providerDisabledTitle(type: ProviderType): string {
-  if (isWiredOAuthProvider(type)) {
-    return '请在 OAuth 分类完成账号登录；登录成功后会自动出现在模型连接。';
-  }
-  return '该账号登录暂未接入聊天发送；当前请使用同一家厂商的模型密钥。';
-}
-
-function providerDisabledAriaLabel(type: ProviderType, name: string): string {
-  if (isWiredOAuthProvider(type)) return `${name}（请从 OAuth 分类登录）`;
-  return `${name}（账号登录暂未接入聊天发送）`;
-}
-
-function providerCatalogAriaLabel(display: ReturnType<typeof providerDisplay>, count: number): string {
-  const parts = [`添加模型供应商：${display.name}`];
-  if (display.badge) parts.push(`标签：${display.badge}`);
-  parts.push(display.description.replace(/[。.!！？?]+$/u, ''));
-  if (count > 0) parts.push(`已配置 ${count} 个`);
-  return parts.join('，');
-}
-
-function isWiredOAuthProvider(type: ProviderType): boolean {
-  return type === 'claude-subscription' || type === 'codex-subscription';
-}
-
-export function ProviderLogo(props: { type: ProviderType; compact?: boolean }) {
-  return (
-    <span className="providerLogo" data-provider={props.type} data-compact={props.compact ? 'true' : undefined} aria-hidden="true">
-      <ProviderLogoMark type={props.type} />
-    </span>
-  );
-}
-
-/**
- * PR-MODEL-OAUTH-SECTION-0 / PR-MODEL-OAUTH-ALL-0 / PR-CLAUDE-CARD-MOVE-0:
- *
- * OAuth login catalog for Settings → 模型. It is rendered by the
- * same tab switcher as 国内 / 海外 / 本地, not as a standalone section
- * pinned above the provider market. All account providers render as
- * equal-size cards; richer provider-specific controls live in the
- * modal opened from that card.
- */
-type OAuthCardId = 'claude' | 'codex' | 'antigravity' | 'cursor';
-type OAuthServiceId = OAuthCardId;
-type BrowserOAuthServiceId = Exclude<OAuthServiceId, 'claude'>;
-
-interface ModelOAuthCard {
-  id: OAuthCardId;
-  providerType: ProviderType;
-  name: string;
-  description: string;
-  status: 'available';
-  statusLabel: string;
-}
-
-const MODEL_OAUTH_CARDS: ReadonlyArray<ModelOAuthCard> = [
-  {
-    id: 'claude',
-    providerType: 'claude-subscription',
-    name: 'Claude Code',
-    description: 'Claude Pro / Max 订阅账号登录。',
-    status: 'available',
-    statusLabel: '可用',
-  },
-  {
-    id: 'codex',
-    providerType: 'codex-subscription',
-    name: 'OpenAI Codex',
-    description: 'ChatGPT Plus / Pro 订阅账号登录。',
-    status: 'available',
-    statusLabel: '可用',
-  },
-  {
-    id: 'antigravity',
-    providerType: 'gemini-cli',
-    name: 'Google Antigravity',
-    description: 'Google 账号登录 Gemini。',
-    status: 'available',
-    statusLabel: '预览',
-  },
-  {
-    id: 'cursor',
-    providerType: 'openai-compatible',
-    name: 'Cursor',
-    description: 'Cursor 订阅账号登录。',
-    status: 'available',
-    statusLabel: '可用',
-  },
-];
-
-function ModelOAuthSection(props: { onConnectionsChanged(): Promise<void> }) {
-  const [openModal, setOpenModal] = useState<OAuthServiceId | null>(null);
-  const toast = useToast();
-  const modelOAuthMountedRef = useRef(false);
-  const modelOAuthRefreshTicketRef = useRef(0);
-  // PR-OAUTH-CARD-LIVE-STATE-0 (WAWQAQ msg d79fd115 follow-up):
-  // before this lift the 3 button cards stayed at the static
-  // "可用 / 预览" label even after the user finished the OAuth
-  // flow in the modal — there was no parent re-fetch. We now
-  // track a runtimeState + email per service so each card can
-  // show "已登录" / the account email inline, and we re-fetch
-  // every time the modal closes (success OR cancel — the user
-  // may have logged out from inside the modal).
-  const [cardStates, setCardStates] = useState<Record<OAuthServiceId, SubscriptionSnapshot | null>>({
-    claude: null,
-    codex: null,
-    cursor: null,
-    antigravity: null,
-  });
-  const [cardRefreshError, setCardRefreshError] = useState<string | null>(null);
-
-  async function refreshAllCards() {
-    const ticket = modelOAuthRefreshTicketRef.current + 1;
-    modelOAuthRefreshTicketRef.current = ticket;
-    const results = await Promise.all(
-      MODEL_OAUTH_CARDS.map(async (card) => {
-        try {
-          const snapshot = await getSubscriptionSnapshot(card.id);
-          return { id: card.id, snapshot } as const;
-        } catch (error) {
-          return { id: card.id, error } as const;
-        }
-      }),
-    );
-    if (!modelOAuthMountedRef.current || modelOAuthRefreshTicketRef.current !== ticket) return false;
-    const failures = results.filter((result) => 'error' in result);
-    setCardStates((prev) => {
-      const next = { ...prev };
-      for (const result of results) {
-        if ('snapshot' in result && result.snapshot !== undefined) next[result.id] = result.snapshot;
-      }
-      return next;
-    });
-    if (failures.length > 0) {
-      const firstFailure = failures[0];
-      const message = firstFailure && 'error' in firstFailure
-        ? subscriptionActionErrorMessage(firstFailure.error)
-        : '登录服务暂时不可用，请检查网络后重试。';
-      setCardRefreshError(message);
-      toast.error('刷新 OAuth 登录状态失败', message);
-      return false;
-    }
-    setCardRefreshError(null);
-    return true;
-  }
-
-  async function refreshAfterModalClose() {
-    const refreshed = await refreshAllCards();
-    if (!modelOAuthMountedRef.current || !refreshed) return;
-    try {
-      await props.onConnectionsChanged();
-    } catch (error) {
-      if (!modelOAuthMountedRef.current) return;
-      toast.error('刷新模型连接失败', subscriptionActionErrorMessage(error));
-    }
-  }
-
-  useEffect(() => {
-    modelOAuthMountedRef.current = true;
-    void refreshAllCards();
-    return () => {
-      modelOAuthMountedRef.current = false;
-      modelOAuthRefreshTicketRef.current += 1;
-    };
-  }, []);
-
-  return (
-    <div className="providerOAuthCatalog" aria-label="OAuth 登录" data-provider-category="oauth">
-      {cardRefreshError && (
-        <div className="providerOAuthError" role="alert">
-          OAuth 登录状态暂时没刷新成功，已保留上一次状态。{cardRefreshError}
-        </div>
-      )}
-      <div className="providerOAuthGrid">
-        {MODEL_OAUTH_CARDS.map((card) => {
-          const snapshot = cardStates[card.id];
-          const runtimeState = snapshot?.runtimeState ?? 'unknown';
-          const isLoggedIn =
-            runtimeState === 'authenticated' ||
-            runtimeState === 'refreshing' ||
-            runtimeState === 'quota_unavailable' ||
-            runtimeState === 'provider_rejected';
-          const liveBadge = isLoggedIn ? '已登录' : card.statusLabel;
-          const liveDescription = isLoggedIn && snapshot?.email
-            ? snapshot.email
-            : card.description;
-          return (
-            <Item
-              key={card.id}
-              className="providerCatalogRow providerOAuthCard rounded-none"
-              data-card-id={card.id}
-              data-provider={card.providerType}
-              data-status="ready"
-              data-oauth-status={card.status}
-              data-logged-in={isLoggedIn ? 'true' : undefined}
-              aria-label={providerOAuthAriaLabel(card, liveBadge, liveDescription)}
-              render={<button type="button" onClick={() => setOpenModal(card.id)} />}
-            >
-              <ItemMedia>
-                <ProviderLogo type={card.providerType} />
-              </ItemMedia>
-              <ItemContent>
-                <ItemTitle className="providerCatalogTitle">{card.name}</ItemTitle>
-                <ItemDescription className="providerCatalogDesc providerOAuthCardDescription">{liveDescription}</ItemDescription>
-              </ItemContent>
-              <ItemActions className="providerCatalogActions">
-                <span className="providerCatalogBadge providerOAuthCardBadge">{liveBadge}</span>
-                <ChevronRight className="providerCatalogChevron" size={15} strokeWidth={2} aria-hidden="true" />
-              </ItemActions>
-            </Item>
-          );
-        })}
-      </div>
-      {openModal === 'claude' && (
-        <ClaudeSubscriptionModal
-          onClose={() => {
-            setOpenModal(null);
-            void refreshAfterModalClose();
-          }}
-        />
-      )}
-      {openModal !== null && openModal !== 'claude' && (
-        <SubscriptionLoginModal
-          serviceId={openModal}
-          onClose={() => {
-            setOpenModal(null);
-            // Always re-fetch after the modal closes — the user may
-            // have logged in, logged out, or cancelled.
-            void refreshAfterModalClose();
-          }}
-        />
-      )}
-    </div>
-  );
-}
-
-function providerOAuthAriaLabel(card: ModelOAuthCard, badge: string, description: string): string {
-  return `打开 OAuth 登录：${card.name}，状态：${badge}，${description.replace(/[。.!！？?]+$/u, '')}`;
-}
-
-/**
- * Inline modal that drives a Codex / Cursor / Antigravity OAuth
- * flow against the matching `window.maka.<service>Subscription`
- * bridge. Mirrors the ClaudeSubscriptionCard pattern (Settings →
- * 账号) but does NOT expose a paste-code field — these flows are
- * loopback (Codex / Antigravity) or polling (Cursor) so the
- * browser handoff is enough.
- *
- * Tokens never enter the renderer; this component reads only
- * account-state snapshots returned by getAccountState().
- */
-function ClaudeSubscriptionModal(props: { onClose(): void }) {
-  const dialogRef = useRef<HTMLElement>(null);
-  useModalA11y(dialogRef, props.onClose);
-  useProviderSheetBackgroundInert(dialogRef);
-  return (
-    <div className="providerConfigOverlay" role="presentation" onMouseDown={props.onClose}>
-      <section
-        ref={dialogRef as RefObject<HTMLDivElement>}
-        className="providerConfigSheet"
-        role="dialog"
-        aria-modal="true"
-        aria-label="Claude Code 登录"
-        data-subscription="claude"
-        onMouseDown={(event) => event.stopPropagation()}
-      >
-        <header className="providerConfigHeader">
-          <div>
-            <h3>Claude Code</h3>
-            <p>登录 Claude Pro / Max 后，会同步成模型连接。</p>
-          </div>
-          <Button
-            type="button"
-            variant="ghost"
-            onClick={props.onClose}
-            aria-label="关闭"
-          >
-            ×
-          </Button>
-        </header>
-        <ClaudeSubscriptionCard />
-      </section>
-    </div>
-  );
-}
-
-function SubscriptionLoginModal(props: { serviceId: BrowserOAuthServiceId; onClose(): void }) {
-  const dialogRef = useRef<HTMLElement>(null);
-  useModalA11y(dialogRef, props.onClose);
-  useProviderSheetBackgroundInert(dialogRef);
-  const toast = useToast();
-  const bridge = pickSubscriptionBridge(props.serviceId);
-  const [state, setState] = useState<SubscriptionSnapshot | null>(null);
-  const [authRequestId, setAuthRequestId] = useState<string | null>(null);
-  const [stateHint, setStateHint] = useState<string | null>(null);
-  const [pendingAction, setPendingAction] = useState<BrowserSubscriptionPendingAction | null>(null);
-  const pendingActionRef = useRef<BrowserSubscriptionPendingAction | null>(null);
-  const authRequestIdRef = useRef<string | null>(null);
-  const browserSubscriptionMountedRef = useRef(false);
-  const [errorMessage, setErrorMessage] = useState<string | null>(null);
-  const display = subscriptionDisplay(props.serviceId);
-
-  async function refresh(): Promise<boolean> {
-    try {
-      const next = (await bridge.getAccountState()) as SubscriptionSnapshot;
-      if (!browserSubscriptionMountedRef.current) return false;
-      setState(next);
-      setErrorMessage(null);
-    } catch (error) {
-      if (!browserSubscriptionMountedRef.current) return false;
-      const message = subscriptionActionErrorMessage(error);
-      toast.error('刷新登录状态失败', message);
-      setErrorMessage(message);
-    }
-    return true;
-  }
-
-  useEffect(() => {
-    browserSubscriptionMountedRef.current = true;
-    void refresh();
-    return () => {
-      browserSubscriptionMountedRef.current = false;
-      pendingActionRef.current = null;
-      const pendingAuthRequestId = authRequestIdRef.current;
-      authRequestIdRef.current = null;
-      if (pendingAuthRequestId) void bridge.cancelAuthorization(pendingAuthRequestId);
-    };
-  }, []);
-
-  function beginPendingAction(action: BrowserSubscriptionPendingAction): boolean {
-    if (pendingActionRef.current !== null) return false;
-    pendingActionRef.current = action;
-    setPendingAction(action);
-    return true;
-  }
-
-  function finishPendingAction() {
-    pendingActionRef.current = null;
-    if (browserSubscriptionMountedRef.current) setPendingAction(null);
-  }
-
-  async function startLogin() {
-    if (!beginPendingAction('login')) return;
-    setErrorMessage(null);
-    try {
-      const payload = await bridge.getAuthUrl();
-      if ('ok' in payload) {
-        if (!browserSubscriptionMountedRef.current) return;
-        const failureMessage = payload.ok ? '请稍后再试。' : subscriptionResultMessage(payload.message, '无法开始登录，请稍后再试。');
-        toast.error('无法开始登录', failureMessage);
-        setErrorMessage(failureMessage);
-        return;
-      }
-      authRequestIdRef.current = payload.authRequestId;
-      if (!browserSubscriptionMountedRef.current) {
-        authRequestIdRef.current = null;
-        void bridge.cancelAuthorization(payload.authRequestId);
-        return;
-      }
-      setAuthRequestId(payload.authRequestId);
-      setStateHint(payload.stateHint);
-      const opened = await bridge.openAuthUrl(payload.authRequestId);
-      if (!browserSubscriptionMountedRef.current) return;
-      if (!opened.ok) {
-        const message = subscriptionResultMessage(opened.message, '无法打开浏览器，请稍后重试。');
-        toast.error('无法打开浏览器', message);
-        setErrorMessage(message);
-        void bridge.cancelAuthorization(payload.authRequestId);
-        authRequestIdRef.current = null;
-        setAuthRequestId(null);
-        setStateHint(null);
-        return;
-      }
-      const refreshed = await refresh();
-      if (!browserSubscriptionMountedRef.current || !refreshed) return;
-      // Loopback / polling — wait for the backend to complete.
-      const result = await bridge.completeAuthorization(payload.authRequestId);
-      if (!browserSubscriptionMountedRef.current) return;
-      authRequestIdRef.current = null;
-      setAuthRequestId(null);
-      setStateHint(null);
-      if (result.ok) {
-        toast.success('登录成功', `${display.name} 已绑定本机。`);
-        await refresh();
-      } else {
-        const message = subscriptionResultMessage(result.message, '登录未完成，请重新打开浏览器授权。');
-        toast.error('登录未完成', message);
-        setErrorMessage(message);
-      }
-    } catch (error) {
-      if (!browserSubscriptionMountedRef.current) return;
-      const pendingAuthRequestId = authRequestIdRef.current;
-      authRequestIdRef.current = null;
-      if (pendingAuthRequestId) void bridge.cancelAuthorization(pendingAuthRequestId);
-      setAuthRequestId(null);
-      setStateHint(null);
-      const message = subscriptionActionErrorMessage(error);
-      toast.error('登录失败', message);
-      setErrorMessage(message);
-    } finally {
-      finishPendingAction();
-    }
-  }
-
-  async function logout() {
-    if (!beginPendingAction('logout')) return;
-    try {
-      const ok = await toast.confirm({
-        title: `退出 ${display.name} 登录？`,
-        description: '将删除本机保存的订阅凭据，之后需要重新登录才能继续使用这些 OAuth 模型。',
-        confirmLabel: '退出登录',
-        cancelLabel: '取消',
-        destructive: true,
-      });
-      if (!ok) return;
-      const result = await bridge.logout();
-      if (!browserSubscriptionMountedRef.current) return;
-      if (result.ok) {
-        toast.success('已退出登录', '本地凭据已清除。');
-        await refresh();
-      } else {
-        toast.error('退出失败', subscriptionResultMessage(result.message, '退出登录失败，请稍后重试。'));
-      }
-    } catch (error) {
-      if (!browserSubscriptionMountedRef.current) return;
-      toast.error('退出失败', subscriptionActionErrorMessage(error));
-    } finally {
-      finishPendingAction();
-    }
-  }
-
-  const runtimeState = state?.runtimeState ?? 'loading';
-  const isLoggedIn = runtimeState === 'authenticated' || runtimeState === 'refreshing';
-  const actionBusy = pendingAction !== null;
-
-  return (
-    <div className="providerConfigOverlay" role="presentation" onMouseDown={props.onClose}>
-      <section
-        ref={dialogRef as RefObject<HTMLDivElement>}
-        className="providerConfigSheet"
-        role="dialog"
-        aria-modal="true"
-        aria-label={`${display.name} 登录`}
-        data-subscription={props.serviceId}
-        onMouseDown={(event) => event.stopPropagation()}
-      >
-        <header className="providerConfigHeader">
-          <div>
-            <h3>{display.name}</h3>
-            <p>{display.detail}</p>
-          </div>
-          <Button
-            type="button"
-            variant="ghost"
-            onClick={props.onClose}
-            aria-label="关闭"
-          >
-            ×
-          </Button>
-        </header>
-        <div className="settingsConnectionRow" data-status={runtimeState}>
-          <p className="settingsConnectionDetail">
-            {presentSnapshotDetail(state, display)}
-          </p>
-          {stateHint && (
-            <small>提示：state 以 <code>{stateHint}</code> 开头。</small>
-          )}
-          {errorMessage && (
-            <small className="settingsErrorText">{errorMessage}</small>
-          )}
-          <div className="settingsConnectionActions">
-            {!isLoggedIn ? (
-              <Button
-                type="button"
-                onClick={() => void startLogin()}
-                disabled={actionBusy}
-              >
-                {pendingAction === 'login' ? '打开浏览器…' : `登录 ${display.shortName}`}
-              </Button>
-            ) : (
-              <Button
-                type="button"
-                variant="ghost"
-                onClick={() => void logout()}
-                disabled={actionBusy}
-              >
-                {pendingAction === 'logout' ? '退出中…' : '退出登录'}
-              </Button>
-            )}
-          </div>
-        </div>
-      </section>
-    </div>
-  );
-}
-
-type BrowserSubscriptionPendingAction = 'login' | 'logout';
-
-interface SubscriptionSnapshot {
-  runtimeState:
-    | 'not_logged_in'
-    | 'authorizing'
-    | 'authenticated'
-    | 'refreshing'
-    | 'refresh_failed'
-    | 'storage_failed'
-    | 'quota_unavailable'
-    | 'provider_rejected';
-  email?: string;
-  plan?: string;
-  status?: 'preview';
-  errorMessage?: string;
-}
-
-interface SubscriptionBridge {
-  getAuthUrl(): Promise<
-    { authRequestId: string; stateHint: string } | { ok: boolean; reason?: string; message: string }
-  >;
-  openAuthUrl(authRequestId: string): Promise<{ ok: true } | { ok: false; reason: string; message: string }>;
-  completeAuthorization(authRequestId: string): Promise<{ ok: true } | { ok: false; reason: string; message: string }>;
-  cancelAuthorization(authRequestId?: string): Promise<{ ok: true }>;
-  getAccountState(): Promise<unknown>;
-  logout(): Promise<{ ok: true } | { ok: false; reason: string; message: string }>;
-}
-
-function subscriptionActionErrorMessage(error: unknown): string {
-  const message = error instanceof Error
-    ? error.message
-    : typeof error === 'string'
-      ? error
-      : '';
-  return subscriptionResultMessage(message, '登录服务暂时不可用，请检查网络后重试。');
-}
-
-function subscriptionResultMessage(message: string | undefined, fallback: string): string {
-  const raw = redactSecrets(message ?? '').trim();
-  if (!raw) return fallback;
-  const classified = generalizedErrorMessageChinese(new Error(raw), '');
-  if (classified) return classified;
-  return /[\u4e00-\u9fff]/.test(raw) ? raw : fallback;
-}
-
-async function getSubscriptionSnapshot(serviceId: OAuthServiceId): Promise<SubscriptionSnapshot> {
-  if (serviceId === 'claude') {
-    const state = await window.maka.claudeSubscription.getAccountState();
-    return {
-      runtimeState: state.runtimeState,
-      email: state.profile?.email,
-      errorMessage: state.errorMessage,
-    };
-  }
-  return (await pickSubscriptionBridge(serviceId).getAccountState()) as SubscriptionSnapshot;
-}
-
-function pickSubscriptionBridge(serviceId: BrowserOAuthServiceId): SubscriptionBridge {
-  switch (serviceId) {
-    case 'codex':
-      return window.maka.codexSubscription as unknown as SubscriptionBridge;
-    case 'cursor':
-      return window.maka.cursorSubscription as unknown as SubscriptionBridge;
-    case 'antigravity':
-      return window.maka.antigravitySubscription as unknown as SubscriptionBridge;
-  }
-}
-
-interface SubscriptionDisplay {
-  name: string;
-  shortName: string;
-  detail: string;
-}
-
-function subscriptionDisplay(serviceId: BrowserOAuthServiceId): SubscriptionDisplay {
-  switch (serviceId) {
-    case 'codex':
-      return {
-        name: 'OpenAI Codex',
-        shortName: 'Codex',
-        detail: '点击下方按钮打开浏览器登录，授权完成后会自动回写到本机（127.0.0.1:1455）。',
-      };
-    case 'cursor':
-      return {
-        name: 'Cursor',
-        shortName: 'Cursor',
-        detail: '点击下方按钮打开浏览器登录；Maka 会自动等待 Cursor 后端确认凭据。',
-      };
-    case 'antigravity':
-      return {
-        name: 'Google Antigravity',
-        shortName: 'Antigravity',
-        // OAuth flow + token persistence + IPC handlers ARE wired
-        // and tested; the only thing gating real login is the
-        // Google client_id constant (no public upstream plugin source
-        // exposes it). When the user clicks 登录 the service surfaces
-        // that exact reason via its envelope, so this card-level
-        // copy stays factual without claiming the whole thing is
-        // unimplemented.
-        detail: '使用 Google 账号登录给 Gemini 模型。当前为预览状态：需要 Google client_id 后才能完成登录。',
-      };
-  }
-  const _exhaustive: never = serviceId;
-  return _exhaustive;
-}
-
-function presentSnapshotDetail(state: SubscriptionSnapshot | null, display: SubscriptionDisplay): string {
-  if (!state) return '正在加载账号状态…';
-  switch (state.runtimeState) {
-    case 'not_logged_in':
-      return `${display.name} 尚未登录。`;
-    case 'authorizing':
-      return '请在弹出的浏览器窗口完成登录。';
-    case 'authenticated': {
-      const parts = ['已登录'];
-      if (state.email) parts.push(state.email);
-      if (state.plan) parts.push(state.plan);
-      return parts.join(' · ');
-    }
-    case 'refreshing':
-      return '正在刷新访问令牌…';
-    case 'refresh_failed':
-      return subscriptionResultMessage(state.errorMessage, '令牌刷新失败，请重新登录。');
-    case 'storage_failed':
-      return subscriptionResultMessage(state.errorMessage, `${display.name} 本地凭据读取失败，请重新登录。`);
-    case 'quota_unavailable':
-    case 'provider_rejected':
-      return subscriptionResultMessage(state.errorMessage, `${display.name} 已登录，但当前 provider 状态不可用。`);
-  }
-  const _exhaustive: never = state.runtimeState;
-  return _exhaustive;
-}
-
-// Renders the official brand mark (vendored in `provider-brand-marks.tsx`).
-// Kept as a thin wrapper so the many `ProviderLogo` call sites stay put.
-function ProviderLogoMark({ type }: { type: ProviderType }) {
-  return <ProviderBrandMark type={type} />;
-}
-
-function AddProviderForm(props: {
-  bridge: ConnectionsBridge;
-  providerType: ProviderType;
-  existingSlugs: string[];
-  onCancel(): void;
-  onCreated(slug: string): Promise<void>;
-}) {
-  const defaults = PROVIDER_DEFAULTS[props.providerType];
-  const display = providerDisplay(props.providerType);
-  const recommendedDefaultModel = buildCatalogRecommendedDefaultModel(props.providerType);
-  const [slug, setSlug] = useState(() => nextSlug(props.providerType, props.existingSlugs));
-  const [name, setName] = useState(display.name);
-  const [baseUrl, setBaseUrl] = useState(defaults.baseUrl);
-  const [defaultModel, setDefaultModel] = useState(recommendedDefaultModel);
-  const [error, setError] = useState<string | null>(null);
-  const [busy, setBusy] = useState(false);
-  const busyRef = useRef(false);
-  const addProviderMountedRef = useRef(false);
-
-  const requiresBaseUrl = !defaults.baseUrl;
-  const isExperimental = defaults.status === 'phase3-experimental';
-  const isWiredOAuth = isWiredOAuthProvider(props.providerType);
-
-  useEffect(() => {
-    addProviderMountedRef.current = true;
-    return () => {
-      addProviderMountedRef.current = false;
-      busyRef.current = false;
-    };
-  }, []);
-
-  async function submit() {
-    if (busyRef.current) return;
-    setError(null);
-    const slugError = validateSlug(slug);
-    if (slugError) return setError(slugError);
-    if (props.existingSlugs.includes(slug)) return setError('连接标识已存在');
-    if (requiresBaseUrl && !baseUrl.trim()) return setError('这个供应商需要填写服务地址');
-    if (isExperimental) {
-      return setError(isWiredOAuth
-        ? '请到 OAuth 分类完成账号登录；登录成功后会自动创建模型连接。'
-        : '该账号登录暂未接入聊天发送；请先使用同一家厂商的模型密钥。');
-    }
-    busyRef.current = true;
-    setBusy(true);
-    try {
-      const connection = await props.bridge.create({
-        slug,
-        name: name || display.name,
-        providerType: props.providerType,
-        baseUrl: baseUrl || undefined,
-        defaultModel,
-      });
-      if (!addProviderMountedRef.current) return;
-      await props.onCreated(connection.slug);
-    } catch (err) {
-      if (addProviderMountedRef.current) setError(providerPanelActionErrorMessage(err));
-    } finally {
-      busyRef.current = false;
-      if (addProviderMountedRef.current) setBusy(false);
-    }
-  }
-
-  return (
-    <div className="providerEditor">
-      <header>
-        <div>
-          <h3>{isExperimental && isWiredOAuth
-            ? `${display.name} 通过 OAuth 登录`
-            : isExperimental ? '账号登录暂未接入聊天发送' : `添加 ${display.name}`}</h3>
-          <p>{display.description}</p>
-        </div>
-        <span className="settingsBadge">{categoryLabel(defaults.category)}</span>
-      </header>
-      {isExperimental && (
-        <div className="providerUnavailableNotice">
-          <strong>{isWiredOAuth ? '使用 OAuth 分类登录' : '账号登录暂未接入'}</strong>
-          <span>{isWiredOAuth
-            ? '不要在这里手动添加；请回到 OAuth 分类完成登录，Maka 会自动创建并刷新模型连接。'
-            : '这类账号登录暂未接入聊天发送。当前请先使用同一家厂商的模型密钥。'}</span>
-        </div>
-      )}
-      <label>
-        <span>连接标识</span>
-        <Input value={slug} onChange={(event) => setSlug(event.currentTarget.value)} placeholder="my-provider" disabled={isExperimental || busy} aria-label="模型供应商连接标识" />
-      </label>
-      <label>
-        <span>显示名称</span>
-        <Input value={name} onChange={(event) => setName(event.currentTarget.value)} placeholder={display.name} disabled={isExperimental || busy} aria-label="模型供应商显示名称" />
-      </label>
-      <label>
-        <span>服务地址 {requiresBaseUrl ? '（必填）' : ''}</span>
-        <Input
-          value={baseUrl}
-          onChange={(event) => setBaseUrl(event.currentTarget.value)}
-          placeholder={defaults.baseUrl || 'https://…'}
-          disabled={isExperimental || busy}
-          aria-label="模型供应商服务地址"
-        />
-      </label>
-      <label>
-        <span>默认模型</span>
-        <Input
-          value={defaultModel}
-          onChange={(event) => setDefaultModel(event.currentTarget.value)}
-          placeholder={recommendedDefaultModel || 'model-id'}
-          disabled={isExperimental || busy}
-          aria-label="模型供应商默认模型"
-        />
-      </label>
-      {error && <p className="providerError">{error}</p>}
-      <div className="providerActions">
-        <Button variant="ghost" type="button" disabled={busy} onClick={props.onCancel}>取消</Button>
-        <Button type="button" disabled={busy || isExperimental} onClick={submit}>
-          {busy ? '保存中…' : '保存供应商'}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-function ConnectionDetail(props: {
-  bridge: ConnectionsBridge;
-  connection: LlmConnection;
-  isDefault: boolean;
-  onChanged(): Promise<void>;
-  onDeleted(): Promise<void>;
-}) {
-  const { connection } = props;
-  const defaults = PROVIDER_DEFAULTS[connection.providerType];
-  const display = providerDisplay(connection.providerType);
-  const [apiKey, setApiKey] = useState('');
-  const [hasSecret, setHasSecret] = useState<CredentialPresenceStatus>(
-    defaults.authKind === 'none' ? true : 'loading',
-  );
-  const [baseUrl, setBaseUrl] = useState(connection.baseUrl ?? defaults.baseUrl ?? '');
-  const [defaultModel, setDefaultModel] = useState(connection.defaultModel);
-  const [models, setModels] = useState<ModelInfo[]>(connection.models ?? []);
-  // Backend persists the model-list source alongside the model cache, so a
-  // Settings restart no longer has to infer "fetched" from a non-empty array.
-  // A successful provider response may legitimately contain 0 models; source
-  // and length remain separate facts.
-  const [modelSource, setModelSource] = useState<'fetched' | 'fallback'>(
-    connection.modelSource ?? 'fallback',
-  );
-  const syncedConnectionSnapshotRef = useRef(connectionDetailSnapshot(connection, defaults.baseUrl));
-  const [busy, setBusy] = useState(false);
-  const [testing, setTesting] = useState(false);
-  const [fetchingModels, setFetchingModels] = useState(false);
-  const [settingDefault, setSettingDefault] = useState(false);
-  const [deleting, setDeleting] = useState(false);
-  const busyRef = useRef(false);
-  const testingRef = useRef(false);
-  const fetchingModelsRef = useRef(false);
-  const settingDefaultRef = useRef(false);
-  const deletingRef = useRef(false);
-  const connectionDetailMountedRef = useRef(false);
-  const connectionDetailLifecycleRef = useRef(0);
-  const toast = useToast();
-  const needsApiKey = defaults.authKind === 'api_key';
-  const needsOAuth = defaults.authKind === 'oauth_token';
-  const hasFixedOAuthBaseUrl = needsOAuth && Boolean(defaults.baseUrl);
-  const requiresCredential = defaults.authKind !== 'none';
-  const credentialProbePending = requiresCredential && (hasSecret === 'loading' || hasSecret === 'error');
-  const hasUsableCredential = !requiresCredential || hasSecret === true;
-  const credentialTroubleshootingCopy = needsOAuth
-    ? 'OAuth 登录 / 代理设置'
-    : '模型密钥 / 服务地址 / 代理设置';
-  const savedBaseUrl = connection.baseUrl ?? defaults.baseUrl;
-  const draftBaseUrl = hasFixedOAuthBaseUrl ? defaults.baseUrl : baseUrl;
-  const hasSaveChanges =
-    apiKey.length > 0 ||
-    draftBaseUrl !== savedBaseUrl ||
-    defaultModel !== connection.defaultModel;
-  const detailActionBusy = busy || testing || fetchingModels || settingDefault || deleting;
-
-  useEffect(() => {
-    connectionDetailMountedRef.current = true;
-    connectionDetailLifecycleRef.current += 1;
-    return () => {
-      connectionDetailMountedRef.current = false;
-      connectionDetailLifecycleRef.current += 1;
-      busyRef.current = false;
-      testingRef.current = false;
-      fetchingModelsRef.current = false;
-      settingDefaultRef.current = false;
-      deletingRef.current = false;
-    };
-  }, [connection.slug]);
-
-  function isConnectionDetailCurrent(lifecycle: number): boolean {
-    return connectionDetailMountedRef.current && connectionDetailLifecycleRef.current === lifecycle;
-  }
-
-  useEffect(() => {
-    const lifecycle = connectionDetailLifecycleRef.current;
-    if (defaults.authKind === 'none') {
-      if (isConnectionDetailCurrent(lifecycle)) setHasSecret(true);
-      return;
-    }
-    setHasSecret('loading');
-    void props.bridge
-      .hasSecret(connection.slug)
-      .then((next) => {
-        if (isConnectionDetailCurrent(lifecycle)) setHasSecret(next);
-      })
-      .catch((error) => {
-        if (!isConnectionDetailCurrent(lifecycle)) return;
-        setHasSecret('error');
-        toast.error('读取模型凭据状态失败', providerPanelActionErrorMessage(error));
-      });
-  }, [props.bridge, connection.slug, defaults.authKind, toast]);
-
-  useEffect(() => {
-    const nextSnapshot = connectionDetailSnapshot(connection, defaults.baseUrl);
-    const previousSnapshot = syncedConnectionSnapshotRef.current;
-    const localStillSynced = connectionDetailDraftMatchesSnapshot(
-      { baseUrl, defaultModel, models, modelSource },
-      previousSnapshot,
-    );
-    const localAlreadyMatchesNext = connectionDetailDraftMatchesSnapshot(
-      { baseUrl, defaultModel, models, modelSource },
-      nextSnapshot,
-    );
-
-    if (connection.slug !== previousSnapshot.slug || (apiKey.length === 0 && localStillSynced)) {
-      setBaseUrl(nextSnapshot.baseUrl);
-      setDefaultModel(nextSnapshot.defaultModel);
-      setModels(nextSnapshot.models);
-      setModelSource(nextSnapshot.modelSource);
-      syncedConnectionSnapshotRef.current = nextSnapshot;
-      return;
-    }
-
-    if (localAlreadyMatchesNext) {
-      syncedConnectionSnapshotRef.current = nextSnapshot;
-    }
-  }, [
-    apiKey.length,
-    baseUrl,
-    connection,
-    defaultModel,
-    defaults.baseUrl,
-    modelSource,
-    models,
-  ]);
-
-  // Picker entries come from the same catalog merge path as Chat and Daily
-  // Review, but use the local unsaved editor draft for model/default changes.
-  const modelChoices = buildCatalogModelChoices({
-    slug: connection.slug,
-    providerType: connection.providerType,
-    defaultModel,
-    models: modelSource === 'fetched' || models.length > 0 ? models : undefined,
-    modelSource,
-    modelsFetchedAt: connection.modelsFetchedAt,
-  });
-  const catalogFallbackCount = modelChoices.filter((choice) => choice.source === 'static_catalog').length;
-
-  async function save() {
-    if (busyRef.current || testingRef.current || fetchingModelsRef.current || settingDefaultRef.current || deletingRef.current) return;
-    if (!canSaveDefaultModelChange(connection.defaultModel, defaultModel, modelChoices)) {
-      toast.error(
-        '默认模型不可用',
-        '请选择一个当前可用于聊天的模型后再保存。',
-      );
-      return;
-    }
-    const lifecycle = connectionDetailLifecycleRef.current;
-    busyRef.current = true;
-    setBusy(true);
-    let saved = false;
-    try {
-      await props.bridge.update(connection.slug, {
-        baseUrl: hasFixedOAuthBaseUrl ? defaults.baseUrl : baseUrl || undefined,
-        defaultModel,
-        ...(apiKey ? { apiKey } : {}),
-      });
-      saved = true;
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      const wroteNewKey = apiKey.length > 0;
-      setApiKey('');
-      const nextHasSecret = requiresCredential ? await props.bridge.hasSecret(connection.slug) : true;
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      setHasSecret(nextHasSecret);
-      await props.onChanged();
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      // Auto-fetch live model list as soon as the secret is in place. Without
-      // this, the user lands on a Settings · 模型 row whose `defaultModel`
-      // dropdown only contains the static fallback list (e.g. Z.ai → just
-      // glm-4.7 / 4.6 / 4.5), which looks like Maka doesn't support newer
-      // models. Auto-fetch on save closes that gap.
-      if (nextHasSecret && (wroteNewKey || (!needsApiKey && models.length === 0))) {
-        void refreshModels({ silent: true });
-      }
-    } catch (error) {
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      if (saved && requiresCredential) {
-        setHasSecret('error');
-      }
-      toast.error(
-        saved ? '刷新模型连接失败' : '保存模型连接失败',
-        providerPanelActionErrorMessage(error),
-      );
-    } finally {
-      busyRef.current = false;
-      if (isConnectionDetailCurrent(lifecycle)) setBusy(false);
-    }
-  }
-
-  async function runTest() {
-    if (testingRef.current || busyRef.current || fetchingModelsRef.current || settingDefaultRef.current || deletingRef.current) return;
-    const lifecycle = connectionDetailLifecycleRef.current;
-    testingRef.current = true;
-    setTesting(true);
-    try {
-      const result: ConnectionTestResult = await props.bridge.test(connection.slug, { model: defaultModel });
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      if (result.ok) {
-        toast.success(
-          `连接成功 · ${connection.name}`,
-          `${result.modelTested} · ${result.latencyMs} ms`,
-        );
-      } else {
-        toast.error(
-          `连接失败 · ${connection.name}`,
-          connectionTestFailureMessage(result, credentialTroubleshootingCopy),
-        );
-      }
-    } catch (error) {
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      const message = providerPanelActionErrorMessage(error);
-      toast.error(`连接测试出错 · ${connection.name}`, message);
-    } finally {
-      testingRef.current = false;
-      if (isConnectionDetailCurrent(lifecycle)) setTesting(false);
-    }
-  }
-
-  async function refreshModels(opts: { silent?: boolean } = {}) {
-    if (fetchingModelsRef.current) return;
-    if (!opts.silent && (busyRef.current || testingRef.current || settingDefaultRef.current || deletingRef.current)) return;
-    const lifecycle = connectionDetailLifecycleRef.current;
-    fetchingModelsRef.current = true;
-    setFetchingModels(true);
-    try {
-      // Backend (xuan `81ed044`) returns a `ModelDiscoveryResult` envelope —
-      // `{ models, source: 'fetched' | 'fallback', fetchedAt }` — and throws
-      // a generalizedErrorMessage on failure. We trust `result.source`
-      // verbatim instead of inferring from list length, so a provider that
-      // legitimately returns 0 models still reads as 'fetched'.
-      const result = await props.bridge.fetchModels(connection.slug);
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      setModels(result.models);
-      setModelSource(result.source);
-      await props.onChanged();
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      if (!opts.silent) {
-        toast.success(`已拉取 ${result.models.length} 个模型 · ${connection.name}`);
-      }
-    } catch (error) {
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      const message = providerPanelActionErrorMessage(error);
-      // Leave the previously-known source / models intact (so the dropdown
-      // doesn't suddenly empty out), but downgrade the source label back to
-      // 'fallback' if we have nothing fresh to show — the failed fetch
-      // means whatever's on screen is not from the latest probe.
-      if (models.length === 0) setModelSource('fallback');
-      toast.error(
-        `拉取模型失败 · ${connection.name}`,
-        `${message} · 当前继续显示静态列表，请确认 ${credentialTroubleshootingCopy} 后重试。`,
-      );
-    } finally {
-      fetchingModelsRef.current = false;
-      if (isConnectionDetailCurrent(lifecycle)) setFetchingModels(false);
-    }
-  }
-
-  async function setAsDefault() {
-    if (settingDefaultRef.current || busyRef.current || testingRef.current || fetchingModelsRef.current || deletingRef.current) return;
-    if (!connection.enabled) {
-      toast.error('无法设为默认', '这个模型连接已禁用，请重新登录或启用后再设为默认。');
-      return;
-    }
-    const lifecycle = connectionDetailLifecycleRef.current;
-    settingDefaultRef.current = true;
-    setSettingDefault(true);
-    try {
-      await props.bridge.setDefault(connection.slug);
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      await props.onChanged();
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      toast.success(`已设为默认 · ${connection.name}`);
-    } catch (error) {
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      toast.error('切换默认失败', providerPanelActionErrorMessage(error));
-    } finally {
-      settingDefaultRef.current = false;
-      if (isConnectionDetailCurrent(lifecycle)) setSettingDefault(false);
-    }
-  }
-
-  async function remove() {
-    if (deletingRef.current || busyRef.current || testingRef.current || fetchingModelsRef.current || settingDefaultRef.current) return;
-    const lifecycle = connectionDetailLifecycleRef.current;
-    deletingRef.current = true;
-    setDeleting(true);
-    const ok = await toast.confirm({
-      title: `删除供应商 ${connection.name}？`,
-      description: '将从模型连接中移除这个供应商配置；如需再次使用，需要重新添加凭据。',
-      confirmLabel: '删除',
-      cancelLabel: '取消',
-      destructive: true,
-    });
-    if (!isConnectionDetailCurrent(lifecycle)) return;
-    if (!ok) {
-      deletingRef.current = false;
-      setDeleting(false);
-      return;
-    }
-    let deleted = false;
-    try {
-      await props.bridge.delete(connection.slug);
-      deleted = true;
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      await props.onDeleted();
-    } catch (error) {
-      if (!isConnectionDetailCurrent(lifecycle)) return;
-      toast.error(
-        deleted ? '刷新模型列表失败' : '删除模型连接失败',
-        providerPanelActionErrorMessage(error),
-      );
-    } finally {
-      deletingRef.current = false;
-      if (isConnectionDetailCurrent(lifecycle)) setDeleting(false);
-    }
-  }
-
-  return (
-    <div className="providerEditor">
-      <header>
-        <div>
-          <h3>{connection.name}</h3>
-          <p>{display.name}</p>
-        </div>
-        <span className="providerHeaderBadges">
-          {props.isDefault && <span className="settingsBadge">默认</span>}
-          <span className="settingsBadge">{categoryLabel(defaults.category)}</span>
-        </span>
-      </header>
-      <FieldRoot className="grid gap-1.5">
-        <Label className="text-xs text-foreground-60">连接标识</Label>
-        <Input value={connection.slug} disabled aria-label="模型连接标识" />
-      </FieldRoot>
-      <FieldRoot className="grid gap-1.5">
-        <Label className="text-xs text-foreground-60">服务地址</Label>
-        {hasFixedOAuthBaseUrl && <FieldDescription>OAuth 固定</FieldDescription>}
-        <Input
-          value={hasFixedOAuthBaseUrl ? defaults.baseUrl : baseUrl}
-          onChange={(event) => setBaseUrl(event.currentTarget.value)}
-          placeholder={defaults.baseUrl}
-          readOnly={hasFixedOAuthBaseUrl}
-          disabled={detailActionBusy}
-          aria-readonly={hasFixedOAuthBaseUrl ? 'true' : undefined}
-          aria-label={hasFixedOAuthBaseUrl ? '模型连接服务地址，OAuth 固定' : '模型连接服务地址'}
-        />
-      </FieldRoot>
-      {needsApiKey && (
-        <FieldRoot className="grid gap-1.5">
-          <Label className="text-xs text-foreground-60">模型密钥</Label>
-          {hasSecret === true && <FieldDescription>已设置，粘贴新值可替换</FieldDescription>}
-          {hasSecret === 'loading' && <FieldDescription>正在读取状态</FieldDescription>}
-          {hasSecret === 'error' && <FieldDescription>凭据状态未知</FieldDescription>}
-          <PasswordInput
-            value={apiKey}
-            onChange={setApiKey}
-            placeholder={hasSecret === true ? '••••••••' : '粘贴模型密钥'}
-            ariaLabel={`${display.name} 模型密钥`}
-            disabled={detailActionBusy}
-          />
-        </FieldRoot>
-      )}
-      {needsOAuth && (
-        <div className="providerUnavailableNotice" data-auth-kind="oauth">
-          <strong>
-            {hasSecret === true
-              ? 'OAuth 已登录'
-              : hasSecret === 'loading'
-                ? 'OAuth 状态读取中'
-                : hasSecret === 'error'
-                  ? 'OAuth 状态未知'
-                  : '等待 OAuth 登录'}
-          </strong>
-          <span>
-            {hasSecret === true
-              ? '该模型连接使用主进程保存的 OAuth access token，不在这里显示或编辑令牌。'
-              : hasSecret === 'loading'
-                ? '正在读取本机 OAuth 登录状态，读取完成前不会把未知状态显示成未登录。'
-                : hasSecret === 'error'
-                  ? '暂时无法读取本机 OAuth 登录状态；请刷新页面或重新打开设置。'
-                  : '请到上方 OAuth 分类完成登录；登录成功后会自动出现在模型连接里。'}
-          </span>
-        </div>
-      )}
-      {credentialProbePending && (
-        <p className="providerError" role="alert">
-          {hasSecret === 'loading'
-            ? '正在读取模型凭据状态，读取完成前暂不测试连接或刷新模型。'
-            : '模型凭据状态暂时没刷新成功，已避免把未知状态显示成未登录或未配置。'}
-        </p>
-      )}
-      <ModelTable
-        modelChoices={modelChoices}
-        defaultModel={defaultModel}
-        onPickDefault={(id) => setDefaultModel(id)}
-        modelSource={modelSource}
-        modelsFetchedAt={connection.modelsFetchedAt}
-        fallbackCount={catalogFallbackCount}
-        canRefresh={!detailActionBusy && hasUsableCredential}
-        fetchingModels={fetchingModels}
-        disabled={detailActionBusy}
-        onRefresh={() => void refreshModels()}
-      />
-      {defaults.signupUrl && (
-        <a className="providerExternalLink" href={defaults.signupUrl} target="_blank" rel="noreferrer noopener">
-          获取模型密钥
-        </a>
-      )}
-      <div className="providerActions">
-        <Button type="button" disabled={detailActionBusy || !hasSaveChanges} onClick={save}>
-          {busy ? '保存中…' : '保存修改'}
-        </Button>
-        <Button variant="secondary" type="button" disabled={detailActionBusy || !hasUsableCredential} onClick={runTest}>
-          {testing ? '测试中…' : '测试连接'}
-        </Button>
-        {!props.isDefault && connection.enabled && (
-          <Button variant="secondary" type="button" disabled={detailActionBusy} onClick={setAsDefault}>
-            {settingDefault ? '设置中…' : '设为默认'}
-          </Button>
-        )}
-        <Button variant="destructive" type="button" disabled={detailActionBusy} onClick={remove}>
-          {deleting ? '删除中…' : '删除'}
-        </Button>
-      </div>
-    </div>
-  );
-}
-
-type ConnectionDetailSnapshot = {
-  slug: string;
-  baseUrl: string;
-  defaultModel: string;
-  models: ModelInfo[];
-  modelSource: 'fetched' | 'fallback';
-};
-
-function connectionDetailSnapshot(
-  connection: LlmConnection,
-  defaultBaseUrl: string | undefined,
-): ConnectionDetailSnapshot {
-  return {
-    slug: connection.slug,
-    baseUrl: connection.baseUrl ?? defaultBaseUrl ?? '',
-    defaultModel: connection.defaultModel,
-    models: connection.models ?? [],
-    modelSource: connection.modelSource ?? 'fallback',
-  };
-}
-
-function connectionDetailDraftMatchesSnapshot(
-  draft: {
-    baseUrl: string;
-    defaultModel: string;
-    models: ModelInfo[];
-    modelSource: 'fetched' | 'fallback';
-  },
-  snapshot: ConnectionDetailSnapshot,
-): boolean {
-  return draft.baseUrl === snapshot.baseUrl &&
-    draft.defaultModel === snapshot.defaultModel &&
-    draft.modelSource === snapshot.modelSource &&
-    modelListsEqual(draft.models, snapshot.models);
-}
-
-function modelListsEqual(left: ModelInfo[], right: ModelInfo[]): boolean {
-  if (left.length !== right.length) return false;
-  for (let index = 0; index < left.length; index += 1) {
-    const leftModel = left[index];
-    const rightModel = right[index];
-    if (leftModel.id !== rightModel.id) return false;
-    if (leftModel.contextWindow !== rightModel.contextWindow) return false;
-    if (leftModel.maxOutputTokens !== rightModel.maxOutputTokens) return false;
-    if (leftModel.capabilities?.chat !== rightModel.capabilities?.chat) return false;
-    if (leftModel.capabilities?.vision !== rightModel.capabilities?.vision) return false;
-    if (leftModel.capabilities?.reasoning !== rightModel.capabilities?.reasoning) return false;
-    if (leftModel.capabilities?.functionCalling !== rightModel.capabilities?.functionCalling) return false;
-    if (leftModel.capabilities?.imageGeneration !== rightModel.capabilities?.imageGeneration) return false;
-  }
-  return true;
-}
-
-/**
- * UI-02 provider model workspace (per @kenji backlog item):
- *
- *   - Source/fetchedAt header (driven by persisted backend metadata)
- *   - Search box to filter long catalogs
- *   - Per-row default radio + capability chips (vision / reasoning /
- *     function calling) when present
- *   - Default model gets a tinted background + "默认" badge
- *   - Empty state distinguishes "fetched 0" from "haven't fetched yet"
- *   - Refresh button anchored to the header
- *
- * Replaces the dropdown + "刷新模型列表" pair the editor used to ship
- * with. The picker is now a workspace, not a form field.
- */
-function ModelTable(props: {
-  modelChoices: ModelCatalogEntry[];
-  defaultModel: string;
-  onPickDefault(id: string): void;
-  modelSource: 'fetched' | 'fallback';
-  modelsFetchedAt?: number;
-  fallbackCount: number;
-  canRefresh: boolean;
-  fetchingModels: boolean;
-  disabled?: boolean;
-  onRefresh(): void;
-}) {
-  const [query, setQuery] = useState('');
-  const listRef = useRef<HTMLUListElement>(null);
-  const filtered = useMemo(() => {
-    const q = query.trim().toLowerCase();
-    if (!q) return props.modelChoices;
-    return props.modelChoices.filter((m) => {
-      const displayName = modelTableDisplayLabel(m).toLowerCase();
-      return m.id.toLowerCase().includes(q) || displayName.includes(q);
-    });
-  }, [props.modelChoices, query]);
-  const selectableModelIds = useMemo(() => selectableDefaultModelIds(filtered), [filtered]);
-  const tabbableModelId = tabbableRadioId(props.defaultModel || undefined, selectableModelIds);
-
-  const headerLine =
-    props.modelSource === 'fetched'
-      ? props.modelChoices.length > 0
-        ? `实时拉取的 ${props.modelChoices.length} 个模型${formatFetchedAtSuffix(props.modelsFetchedAt)}`
-        : '已成功调用供应商接口，但返回 0 个模型 — 该供应商可能未对当前模型密钥开放任何模型。'
-      : `静态备用列表（${props.fallbackCount} 项）。点「刷新模型列表」拉取该供应商的真实模型清单。`;
-
-  // ARIA radiogroup keyboard pattern: arrow keys move focus AND select.
-  // Space/Enter on a focused radio just trigger the native button click.
-  // The pure `nextRadioId` helper is unit-tested in
-  // `apps/desktop/src/main/__tests__/model-table-keyboard.test.ts`.
-  function onListKeyDown(event: ReactKeyboardEvent<HTMLUListElement>) {
-    if (props.disabled) return;
-    const list = listRef.current;
-    if (!list) return;
-    const radios = Array.from(list.querySelectorAll<HTMLButtonElement>('button[role="radio"]'));
-    if (radios.length === 0) return;
-    const focusedRadio = (document.activeElement as HTMLElement | null)?.closest<HTMLButtonElement>('button[role="radio"]');
-    const currentId = focusedRadio?.dataset.modelId;
-    const nextId = nextRadioId(currentId, selectableModelIds, event.key);
-    if (nextId === null || nextId === currentId) return;
-    event.preventDefault();
-    const next = radios.find((radio) => radio.dataset.modelId === nextId);
-    next?.focus({ preventScroll: false });
-    next?.scrollIntoView({ block: 'nearest' });
-    // ARIA radiogroup pattern (per @xuan PR92 follow-up): arrow keys move
-    // focus AND select. Safe because `onPickDefault` updates local form
-    // state only — persistence happens on "保存修改", so scanning models
-    // with the arrow keys doesn't write to disk on every keystroke.
-    props.onPickDefault(nextId);
-  }
-
-  // @kenji PR91 follow-up #2: when search filters out the currently-selected
-  // default, surface a one-line hint so the user doesn't lose track of which
-  // model is in effect. Click the hint to clear the search.
-  const defaultHidden =
-    query.trim().length > 0 &&
-    props.defaultModel.length > 0 &&
-    filtered.every((m) => m.id !== props.defaultModel);
-
-  return (
-    <div className="modelTable" data-source={props.modelSource}>
-      <header className="modelTableHeader">
-        <div className="modelTableHeaderText">
-          <strong>模型</strong>
-          <small>{headerLine}</small>
-          <small className="modelTableStickyHint">
-            默认模型只用于新建会话；已有会话会保留创建时的模型选择。
-          </small>
-        </div>
-        <Button
-          type="button"
-          disabled={!props.canRefresh}
-          onClick={props.onRefresh}
-        >
-          {props.fetchingModels ? '拉取中…' : '刷新模型列表'}
-        </Button>
-      </header>
-
-      {props.modelChoices.length > 6 && (
-        <Input
-          type="search"
-          className="modelTableSearch"
-          placeholder={`在 ${props.modelChoices.length} 个模型中搜索…`}
-          value={query}
-          onChange={(event) => setQuery(event.currentTarget.value)}
-          autoComplete="off"
-          spellCheck={false}
-          aria-label="搜索模型"
-        />
-      )}
-
-      {defaultHidden && (
-        <Button
-          type="button"
-          variant="ghost"
-          className="modelTableDefaultHint"
-          onClick={() => setQuery('')}
-          title="清空搜索"
-        >
-          当前默认 <code>{props.defaultModel}</code> 不在搜索结果中 · 点这里清空搜索
-        </Button>
-      )}
-
-      {props.modelChoices.length === 0 ? (
-        <div className="modelTableEmpty">
-          {props.modelSource === 'fetched'
-            ? '拉取返回 0 个模型。请检查账号方案或重新拉取。'
-            : '尚无模型。点「刷新模型列表」拉取或先配置模型密钥。'}
-        </div>
-      ) : filtered.length === 0 ? (
-        <div className="modelTableEmpty">没有匹配 “{query}” 的模型。</div>
-      ) : (
-        <ul
-          ref={listRef}
-          className="modelTableList"
-          role="radiogroup"
-          aria-label="默认模型"
-          onKeyDown={onListKeyDown}
-        >
-          {filtered.map((model) => {
-            const isDefault = model.id === props.defaultModel;
-            const displayName = modelTableDisplayLabel(model);
-            const showRawId = displayName !== model.id;
-            const warning = modelTableEntryWarning(model);
-            const canPickDefault = canPickDefaultModel(model);
-            return (
-              <li key={model.id} role="none">
-                <Button
-                  type="button"
-                  className="modelTableRow"
-                  variant="ghost"
-                  role="radio"
-                  aria-checked={isDefault}
-                  aria-disabled={!canPickDefault || props.disabled ? true : undefined}
-                  data-default={isDefault ? 'true' : undefined}
-                  data-disabled={!canPickDefault ? 'true' : undefined}
-                  data-model-id={model.id}
-                  disabled={props.disabled || !canPickDefault}
-                  // Only the active radio is in the tab order; arrow keys
-                  // move focus inside the group. Standard ARIA radiogroup.
-                  tabIndex={tabbableModelId === model.id ? 0 : -1}
-                  onClick={() => {
-                    if (!canPickDefault) return;
-                    props.onPickDefault(model.id);
-                  }}
-                >
-                  <span className="modelTableRowRadio" aria-hidden="true" />
-                  <span className="modelTableRowText">
-                    <span className="modelTableRowName">{displayName}</span>
-                    {showRawId && <code className="modelTableRowId">{model.id}</code>}
-                    {warning && <span className="modelTableRowWarning">{warning}</span>}
-                  </span>
-                  <ModelCapabilityChips model={model} />
-                  {isDefault && <span className="modelTableDefaultBadge">默认</span>}
-                </Button>
-              </li>
-            );
-          })}
-        </ul>
-      )}
-    </div>
-  );
-}
-
-function modelTableDisplayLabel(model: Pick<ModelCatalogEntry, 'id' | 'displayName'>): string {
-  return model.displayName?.trim() || model.id;
-}
-
-function modelTableEntryWarning(model: Pick<ModelCatalogEntry, 'unavailableReason' | 'availability'>): string | null {
-  if (model.unavailableReason === 'not_in_live_list') {
-    return '已保存，但当前模型列表未返回，可能不可用。';
-  }
-  if (model.availability === 'blocked') return '当前不可用。';
-  if (model.availability === 'warning') return '模型列表可能已过期。';
-  return null;
-}
-
-function ModelCapabilityChips(props: { model: Pick<ModelCatalogEntry, 'capabilities' | 'contextWindow'> }) {
-  const caps = props.model.capabilities;
-  const chips: string[] = [];
-  if (caps.vision) chips.push('vision');
-  if (caps.reasoning) chips.push('reasoning');
-  if (caps.functionCalling) chips.push('tools');
-  if (props.model.contextWindow) {
-    // 200_000 → "200K", 1_000_000 → "1M". Compact for the row.
-    chips.push(formatContextWindow(props.model.contextWindow));
-  }
-  if (chips.length === 0) return null;
-  return (
-    <span className="modelTableChips">
-      {chips.map((c) => (
-        <span key={c} className="modelTableChip">{c}</span>
-      ))}
-    </span>
-  );
-}
-
-function formatContextWindow(tokens: number): string {
-  if (tokens >= 1_000_000) return `${(tokens / 1_000_000).toFixed(tokens % 1_000_000 === 0 ? 0 : 1)}M ctx`;
-  if (tokens >= 1_000) return `${Math.round(tokens / 1_000)}K ctx`;
-  return `${tokens} ctx`;
-}
-
-export function providerDisplay(type: ProviderType): { name: string; description: string; badge?: string } {
-  switch (type) {
-    // Descriptions stay version-agnostic on purpose: they name the
-    // PROVIDER and how you connect (official key / protocol-compatible /
-    // local), never a specific model generation — model names go stale
-    // (GPT-4o, DeepSeek-V3, …) but the provider and access path do not.
-    case 'anthropic':
-      return { name: 'Anthropic', description: 'Anthropic 官方接入', badge: 'API' };
-    case 'kimi-coding-plan':
-      return { name: 'Kimi Coding Plan', description: '月之暗面 · Anthropic 兼容', badge: 'Coding' };
-    case 'openai':
-      return { name: 'OpenAI', description: 'OpenAI 官方接入', badge: 'API' };
-    case 'google':
-      return { name: 'Google Gemini', description: 'Google AI Studio 接入', badge: 'API' };
-    case 'deepseek':
-      return { name: 'DeepSeek', description: 'DeepSeek 官方接入', badge: 'API' };
-    case 'moonshot':
-      return { name: 'Moonshot', description: 'Moonshot 官方接入', badge: 'API' };
-    case 'zai-coding-plan':
-      return { name: 'Z.AI Coding Plan', description: '智谱 · OpenAI 兼容', badge: 'Coding' };
-    case 'ollama':
-      return { name: 'Ollama', description: '本机运行 · 离线可用', badge: 'Local' };
-    case 'openai-compatible':
-      return { name: 'OpenAI Compatible', description: '中转站、代理服务或自部署网关。', badge: 'Custom' };
-    case 'claude-subscription':
-      return { name: 'Claude Subscription', description: 'Claude Pro / Max 订阅账号登录；登录后自动成为可用模型连接。' };
-    case 'codex-subscription':
-      return { name: 'OpenAI OAuth', description: 'ChatGPT / Codex 账号登录；登录后自动成为可用模型连接。' };
-    case 'gemini-cli':
-      return { name: 'Gemini CLI', description: 'Google 账号登录暂未接入聊天发送。' };
-  }
-}
-
-function categoryLabel(category: ProviderCategory): string {
-  switch (category) {
-    case 'oauth': return 'OAuth';
-    case 'domestic': return '国内';
-    case 'overseas': return '海外';
-    case 'local': return '本地';
-    case 'custom': return 'Custom';
-  }
-}
-
-function nextSlug(type: ProviderType, existing: string[]): string {
-  const base = type.replace(/[^a-z0-9-]/g, '-');
-  if (!existing.includes(base)) return base;
-  for (let i = 2; i < 100; i += 1) {
-    const candidate = `${base}-${i}`;
-    if (!existing.includes(candidate)) return candidate;
-  }
-  return `${base}-${Date.now()}`;
-}
-/**
- * PR-OAUTH-SUBSCRIPTION-0: Claude subscription card.
- *
- * Renders the runtime state, login/logout actions, paste-code modal,
- * and quota meter. Tokens never enter renderer — this component
- * consumes only `SubscriptionAccountState`.
- */
-function ClaudeSubscriptionCard() {
-  const [experimentalEnabled, setExperimentalEnabled] = useState<boolean | null>(null);
-  const [experimentalGateError, setExperimentalGateError] = useState<string | null>(null);
-  const [state, setState] = useState<SubscriptionAccountState | null>(null);
-  const [pendingAction, setPendingAction] = useState<ClaudeSubscriptionPendingAction | null>(null);
-  const pendingActionRef = useRef<ClaudeSubscriptionPendingAction | null>(null);
-  const [authRequestId, setAuthRequestId] = useState<string | null>(null);
-  const claudeAuthRequestIdRef = useRef<string | null>(null);
-  const [stateHint, setStateHint] = useState<string | null>(null);
-  const [pasteValue, setPasteValue] = useState('');
-  const [pasteError, setPasteError] = useState<string | null>(null);
-  const toast = useToast();
-  // PR-FE-BUG-HUNT-1 (kenji bug-hunt 2026-06-24): ClaudeSubscriptionCard
-  // launches a browser OAuth flow that takes seconds-to-minutes to
-  // complete. Closing the Settings modal while a `startLogin` /
-  // `submitPaste` / `logout` / `refreshQuota` call was in flight
-  // would `setState` on an unmounted component (loud warning in dev,
-  // masks real bugs in prod). Mirror the `mountedRef` pattern other
-  // settings sub-cards in this file use.
-  const claudeCardMountedRef = useRef(true);
-  useEffect(() => {
-    claudeCardMountedRef.current = true;
-    return () => {
-      claudeCardMountedRef.current = false;
-      const pendingAuthRequestId = claudeAuthRequestIdRef.current;
-      claudeAuthRequestIdRef.current = null;
-      if (pendingAuthRequestId) void window.maka.claudeSubscription.cancelAuthorization(pendingAuthRequestId);
-    };
-  }, []);
-
-  const refresh = async () => {
-    try {
-      const next = await window.maka.claudeSubscription.getAccountState();
-      if (!claudeCardMountedRef.current) return;
-      setState(next);
-      setPasteError(null);
-    } catch (error) {
-      const message = subscriptionActionErrorMessage(error);
-      if (!claudeCardMountedRef.current) return;
-      toast.error('刷新登录状态失败', message);
-      setPasteError(message);
-    }
-  };
-
-  const refreshExperimentalGate = async () => {
-    try {
-      const flag = await window.maka.claudeSubscription.isExperimentalEnabled();
-      if (!claudeCardMountedRef.current) return;
-      setExperimentalEnabled(flag);
-      setExperimentalGateError(null);
-      if (flag) void refresh();
-    } catch (error) {
-      const message = subscriptionActionErrorMessage(error);
-      if (!claudeCardMountedRef.current) return;
-      setExperimentalEnabled(null);
-      setExperimentalGateError(message);
-      toast.error('读取 Claude 登录开关失败', message);
-    }
-  };
-
-  useEffect(() => {
-    // kenji `1da909d5` blocking concern: Anthropic does not permit
-    // third-party developers to offer Claude.ai login on behalf of
-    // users. Until product/legal sign-off, gate the whole UI behind
-    // `MAKA_CLAUDE_SUBSCRIPTION_EXPERIMENTAL=1`. Loading state also
-    // renders nothing — no teasing UI.
-    let cancelled = false;
-    void window.maka.claudeSubscription
-      .isExperimentalEnabled()
-      .then((flag) => {
-        if (cancelled) return;
-        setExperimentalEnabled(flag);
-        setExperimentalGateError(null);
-        if (flag) void refresh();
-      })
-      .catch((error) => {
-        if (cancelled) return;
-        const message = subscriptionActionErrorMessage(error);
-        setExperimentalEnabled(null);
-        setExperimentalGateError(message);
-        toast.error('读取 Claude 登录开关失败', message);
-      });
-    return () => {
-      cancelled = true;
-    };
-  }, []);
-
-  if (experimentalGateError) {
-    return (
-      <div className="settingsConnectionRow" data-status="error">
-        <div className="settingsConnectionRowHead">
-          <div className="settingsConnectionRowText">
-            <div className="settingsConnectionRowName">
-              <strong>Claude 订阅 (Pro / Max)</strong>
-            </div>
-            <small>无法确认 Claude OAuth 是否可用。没有登录动作会被执行。</small>
-          </div>
-          <span className="settingsConnectionBadge" data-tone="destructive">读取失败</span>
-        </div>
-        <small className="settingsErrorText" role="alert">
-          Claude 登录开关读取失败：{experimentalGateError}
-        </small>
-        <div className="settingsConnectionActions">
-          <Button
-            type="button"
-            onClick={() => void refreshExperimentalGate()}
-          >
-            重试
-          </Button>
-        </div>
-      </div>
-    );
-  }
-
-  if (experimentalEnabled !== true) {
-    return null;
-  }
-
-  function beginPendingAction(action: ClaudeSubscriptionPendingAction): boolean {
-    if (pendingActionRef.current !== null) return false;
-    pendingActionRef.current = action;
-    setPendingAction(action);
-    return true;
-  }
-
-  function finishPendingAction() {
-    pendingActionRef.current = null;
-    setPendingAction(null);
-  }
-
-  async function startLogin() {
-    if (!beginPendingAction('login')) return;
-    try {
-      // kenji `027c93c0` + xuan `2e5be5a`: getAuthUrl now returns
-      // a union — `AuthorizationUrlPayload` on success, or a
-      // `SubscriptionActionResult` envelope when fail-closed
-      // (e.g. experimental flag flipped off after the card
-      // mounted). Discriminate by checking for the `ok` field; the
-      // envelope variant has it, the success payload does not.
-      const payload = await window.maka.claudeSubscription.getAuthUrl();
-      if ('ok' in payload) {
-        if (!claudeCardMountedRef.current) return;
-        // Envelope variant. `ok: true` shouldn't happen for
-        // getAuthUrl (success returns the payload, not an envelope),
-        // so this branch is the failure case in practice.
-        toast.error('无法开始登录', payload.ok ? '请稍后再试。' : subscriptionResultMessage(payload.message, '无法开始登录，请稍后再试。'));
-        return;
-      }
-      claudeAuthRequestIdRef.current = payload.authRequestId;
-      if (!claudeCardMountedRef.current) {
-        claudeAuthRequestIdRef.current = null;
-        void window.maka.claudeSubscription.cancelAuthorization(payload.authRequestId);
-        return;
-      }
-      setAuthRequestId(payload.authRequestId);
-      setStateHint(payload.stateHint);
-      setPasteValue('');
-      setPasteError(null);
-      // kenji `1da909d5` hardening: pass the opaque authRequestId,
-      // NOT the URL. Main looks up the URL it generated.
-      const opened = await window.maka.claudeSubscription.openAuthUrl(payload.authRequestId);
-      if (!claudeCardMountedRef.current) return;
-      if (!opened.ok) {
-        toast.error('无法打开浏览器', subscriptionResultMessage(opened.message, '无法打开浏览器，请稍后重试。'));
-        claudeAuthRequestIdRef.current = null;
-        void window.maka.claudeSubscription.cancelAuthorization(payload.authRequestId);
-        setAuthRequestId(null);
-        setStateHint(null);
-      }
-      await refresh();
-    } catch (error) {
-      const pendingAuthRequestId = claudeAuthRequestIdRef.current;
-      claudeAuthRequestIdRef.current = null;
-      if (pendingAuthRequestId) void window.maka.claudeSubscription.cancelAuthorization(pendingAuthRequestId);
-      const message = subscriptionActionErrorMessage(error);
-      if (!claudeCardMountedRef.current) return;
-      setAuthRequestId(null);
-      setStateHint(null);
-      toast.error('无法开始登录', message);
-      setPasteError(message);
-    } finally {
-      if (claudeCardMountedRef.current) finishPendingAction();
-    }
-  }
-
-  async function submitPaste() {
-    if (!authRequestId) return;
-    if (!beginPendingAction('submit')) return;
-    setPasteError(null);
-    try {
-      const result = await window.maka.claudeSubscription.completeAuthorization(
-        authRequestId,
-        pasteValue,
-      );
-      if (!claudeCardMountedRef.current) return;
-      if (result.ok) {
-        toast.success('登录成功', '已绑定 Claude 订阅。');
-        claudeAuthRequestIdRef.current = null;
-        setAuthRequestId(null);
-        setStateHint(null);
-        setPasteValue('');
-        await refresh();
-      } else {
-        setPasteError(subscriptionResultMessage(result.message, '授权码提交失败，请重新登录后再试。'));
-      }
-    } catch (error) {
-      const message = subscriptionActionErrorMessage(error);
-      if (!claudeCardMountedRef.current) return;
-      toast.error('授权码提交失败', message);
-      setPasteError(message);
-    } finally {
-      if (claudeCardMountedRef.current) finishPendingAction();
-    }
-  }
-
-  async function cancelLogin() {
-    if (!authRequestId) return;
-    if (!beginPendingAction('cancel')) return;
-    try {
-      await window.maka.claudeSubscription.cancelAuthorization(authRequestId);
-      if (!claudeCardMountedRef.current) return;
-      claudeAuthRequestIdRef.current = null;
-      setAuthRequestId(null);
-      setStateHint(null);
-      setPasteValue('');
-      setPasteError(null);
-      await refresh();
-    } catch (error) {
-      if (!claudeCardMountedRef.current) return;
-      toast.error('取消登录失败', subscriptionActionErrorMessage(error));
-    } finally {
-      if (claudeCardMountedRef.current) finishPendingAction();
-    }
-  }
-
-  async function logout() {
-    if (!beginPendingAction('logout')) return;
-    try {
-      const ok = await toast.confirm({
-        title: '退出 Claude Code 登录？',
-        description: '将删除本机保存的订阅凭据，之后需要重新登录才能继续使用 Claude OAuth 模型。',
-        confirmLabel: '退出登录',
-        cancelLabel: '取消',
-        destructive: true,
-      });
-      if (!ok) return;
-      const result = await window.maka.claudeSubscription.logout();
-      if (!claudeCardMountedRef.current) return;
-      if (result.ok) {
-        toast.success('已退出登录', '本地凭据已清除。');
-        await refresh();
-      } else {
-        toast.error('退出失败', subscriptionResultMessage(result.message, '退出登录失败，请稍后重试。'));
-      }
-    } catch (error) {
-      if (!claudeCardMountedRef.current) return;
-      toast.error('退出失败', subscriptionActionErrorMessage(error));
-    } finally {
-      if (claudeCardMountedRef.current) finishPendingAction();
-    }
-  }
-
-  async function refreshQuota() {
-    if (!beginPendingAction('quota')) return;
-    try {
-      await window.maka.claudeSubscription.refreshQuota();
-      if (!claudeCardMountedRef.current) return;
-      await refresh();
-    } catch (error) {
-      if (!claudeCardMountedRef.current) return;
-      toast.error('刷新配额失败', subscriptionActionErrorMessage(error));
-    } finally {
-      if (claudeCardMountedRef.current) finishPendingAction();
-    }
-  }
-
-  // Closed-state render mapping per the runtime state enum.
-  const presentation = state ? presentSubscriptionState(state) : { label: '加载中…', tone: 'muted', detail: '' };
-  const canStartClaudeLogin =
-    state?.runtimeState === 'not_logged_in' ||
-    state?.runtimeState === 'refresh_failed' ||
-    state?.runtimeState === 'storage_failed';
-  const claudeLoginPending = authRequestId !== null || state?.runtimeState === 'authorizing';
-  const actionBusy = pendingAction !== null;
-
-  return (
-    <>
-    <h3 className="settingsSubheading">订阅</h3>
-    <div className="settingsConnectionRow" data-status={state?.runtimeState ?? 'loading'}>
-      <div className="settingsConnectionRowHead">
-        <div className="settingsConnectionRowText">
-          <div className="settingsConnectionRowName">
-            <strong>Claude 订阅 (Pro / Max)</strong>
-          </div>
-          <small>
-            通过 Anthropic 官方 OAuth 登录使用订阅配额。
-            {state?.profile?.email ? ` · ${state.profile.email}` : ''}
-          </small>
-        </div>
-        <span className="settingsConnectionBadge" data-tone={presentation.tone}>
-          {presentation.label}
-        </span>
-      </div>
-      <p className="settingsConnectionDetail">{presentation.detail}</p>
-      {pasteError && !authRequestId && (
-        <small className="settingsErrorText" role="alert">{pasteError}</small>
-      )}
-
-      {state?.quota && (state.quota.fiveHour || state.quota.sevenDay) && (
-        <div className="settingsQuotaSection">
-          {state.quota.fiveHour && (
-            <div className="settingsQuotaRow">
-              <span>5 小时窗口</span>
-              <span>{state.quota.fiveHour.utilization}%</span>
-            </div>
-          )}
-          {state.quota.sevenDay && (
-            <div className="settingsQuotaRow">
-              <span>7 天窗口</span>
-              <span>{state.quota.sevenDay.utilization}%</span>
-            </div>
-          )}
-          <small className="settingsHelpText">
-            数据更新于 <RelativeTime ts={state.quota.fetchedAt} className="settingsHelpInlineTime" />
-          </small>
-        </div>
-      )}
-
-      <div className="settingsConnectionActions">
-        {canStartClaudeLogin || claudeLoginPending ? (
-          <Button
-            type="button"
-            onClick={() => void startLogin()}
-            disabled={actionBusy || claudeLoginPending}
-          >
-            {pendingAction === 'login'
-              ? '打开浏览器…'
-              : claudeLoginPending
-              ? '登录中…'
-              : state?.runtimeState === 'refresh_failed' || state?.runtimeState === 'storage_failed'
-                ? '重新登录'
-                : '登录订阅'}
-          </Button>
-        ) : (
-          <>
-            <Button
-              type="button"
-              onClick={() => void refreshQuota()}
-              disabled={actionBusy}
-            >
-              {pendingAction === 'quota' ? '刷新中…' : '刷新配额'}
-            </Button>
-            <Button
-              type="button"
-              variant="ghost"
-              onClick={() => void logout()}
-              disabled={actionBusy}
-            >
-              {pendingAction === 'logout' ? '退出中…' : '退出登录'}
-            </Button>
-          </>
-        )}
-      </div>
-
-      {authRequestId && (
-        <div className="settingsOauthPastePanel" role="region" aria-label="粘贴授权码">
-          <p>
-            在 Claude.ai 完成登录后，会跳转到 Anthropic 控制台显示一段授权码（含 <code>#</code> 分隔符），
-            把它粘贴到下面：
-          </p>
-          {stateHint && (
-            <small>提示：你的 state 以 <code>{stateHint}</code> 开头。</small>
-          )}
-          <Textarea
-            value={pasteValue}
-            onChange={(event) => setPasteValue(event.currentTarget.value)}
-            placeholder="粘贴授权码（格式：xxx#yyy）"
-            aria-label="授权码"
-            rows={3}
-            spellCheck={false}
-            autoComplete="off"
-          />
-          {pasteError && <small className="settingsErrorText">{pasteError}</small>}
-          <div className="settingsConnectionActions">
-            <Button
-              type="button"
-              onClick={() => void submitPaste()}
-              disabled={actionBusy || pasteValue.trim().length === 0}
-            >
-              {pendingAction === 'submit' ? '提交中…' : '提交授权码'}
-            </Button>
-            <Button
-              type="button"
-              variant="ghost"
-              onClick={() => void cancelLogin()}
-              disabled={actionBusy}
-            >
-              {pendingAction === 'cancel' ? '取消中…' : '取消'}
-            </Button>
-          </div>
-        </div>
-      )}
-    </div>
-    </>
-  );
-}
-
-type ClaudeSubscriptionPendingAction = 'login' | 'submit' | 'cancel' | 'logout' | 'quota';
-
-interface SubscriptionStatePresentation {
-  label: string;
-  tone: string;
-  detail: string;
-}
-
-function presentSubscriptionState(state: SubscriptionAccountState): SubscriptionStatePresentation {
-  switch (state.runtimeState) {
-    case 'not_logged_in':
-      return { label: '未登录', tone: 'muted', detail: '使用 Claude 订阅配额前需要先登录。' };
-    case 'authorizing':
-      return { label: '登录中…', tone: 'info', detail: '请在弹出的浏览器窗口完成登录并粘贴授权码。' };
-    case 'authenticated':
-      return {
-        label: '已登录',
-        tone: 'success',
-        detail: '已绑定 Claude 订阅，并会同步到“模型连接”。',
-      };
-    case 'refreshing':
-      return { label: '刷新中…', tone: 'info', detail: '正在刷新访问令牌。' };
-    case 'refresh_failed':
-      return {
-        label: '刷新失败',
-        tone: 'warning',
-        detail: subscriptionResultMessage(state.errorMessage, '令牌刷新失败，请重新登录。'),
-      };
-    case 'storage_failed':
-      return {
-        label: '凭据读取失败',
-        tone: 'warning',
-        detail: subscriptionResultMessage(state.errorMessage, '本地 OAuth 凭据读取失败，请重新登录。'),
-      };
-    case 'quota_unavailable':
-      return {
-        label: '等待获取配额',
-        tone: 'warning',
-        detail: subscriptionResultMessage(state.errorMessage, '已登录；配额接口当前没有返回可用数据。'),
-      };
-    case 'provider_rejected':
-      return {
-        label: '订阅 API 拒绝',
-        tone: 'destructive',
-        detail: subscriptionResultMessage(state.errorMessage, '订阅端点拒绝了请求，可能需要重新登录。'),
-      };
-    default:
-      return { label: '未知状态', tone: 'muted', detail: '' };
-  }
 }
